add paths for document uploads, success ui, more robust api
