+++ terraform/environments/dev/terraform.tfvars
@@ -1,4 +1,4 @@
 s3_bucket_name          = "sdc-uploads"
-criminal_check_folder   = "criminal_check"
-income_proof_folder     = "income_proof"
-curriculum_plans_folder = "curriculum_plans"
+criminal_check_folder   = "criminal-check"
+income_proof_folder     = "income-proof"
+curriculum_plans_folder = "curriculum-plans"
